refactor(synthdef)!: @sc-app/synthdef-compiler over the wasm build
The package becomes the thin layer over the crate's wasm-bindgen build:
component.ts (explicit init via pkg/init.js, node bytes / browser fetch)
re-exports the SynthDef class, typed parseScgf (SynthDefJson interfaces),
operator lookups and the env build/encode functions; registry.ts caches
ONE registryJson() call and normalizes to the historical shapes
(lowercase rates, {name, default} records); env-registry.ts serves
ENV_SHAPES from envShapesJson with per-entry buildRun over the crate;
ugen-input.ts's k/u/uo/ugenIndex/outputIndex speak the serde shape
({ constant } | { ugen } | { ugenOutput }); builders.ts re-exports the
548 generated typed builder fns. DELETED: the TS specs (the registry
data lives once, in the crate), the TS builders/encoder/env modules,
and the whole sugar/ DSL (fn.toString parsing — unportable; its only
consumer, scopeTapSynthDef, is rewritten imperatively over the SynthDef
class + inAr/scopeOut2Ar builders).

App sweep: compileSynthDef keeps its logic (nodeRate rename, ugenIndex
helper instead of direct tag reads, operator lookups return undefined),
functions.ts calls entry.buildRun, envValue.ts calls encodeEnvRun. Env
constants now carry wire (float32) precision — six test expectations
pin it via Math.fround, matching the server-commands convention. The
package suite re-pins registry content (incl. the SCDoc-reconcile
entries), graph building, operators and env runs against the real wasm;
the migration-window registry-parity script and the TS reconcile script
retire with the TS specs they compared.

+++ src/sc-elements/state/__tests__/arrays.test.ts
@@ -128,7 +128,7 @@ describe("array slot lenses (numeric-tail binds)", () => {
     const env = host.querySelector('sc-var[name="env"]') as ScVar;
     const sustain = host.querySelector('sc-var[name="sustain"]') as ScVar;
     expect(env._state).toHaveLength(36); // the static call literal evaluated
-    expect(sustain._state).toBe(0.7); // slot 8 = the adsr sustain level
+    expect(sustain._state).toBe(Math.fround(0.7)); // slot 8 (wire f32)
 
     // The WRITE half: a click commits 0.42 through the env.9 lens — one
     // fresh array, one statechange, the other lens recomputes.
@@ -137,7 +137,7 @@ describe("array slot lenses (numeric-tail binds)", () => {
     const after = env._state as number[];
     expect(after[9]).toBe(0.42);
     expect(after).not.toBe(before); // immutable-by-identity
-    expect(after[8]).toBe(0.7); // untouched slots survive
+    expect(after[8]).toBe(Math.fround(0.7)); // untouched slots survive (wire f32)
 
     // The READ half stays live: an editor-style whole-array write moves the
     // lens value.
